[visquv] Rename folder to visquv
This will rename folder name to visquv as visq + u + viewer.

ONE-vscode-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
